feat(267): shared spatial payload for map and alerts; MapView FeatureCollection only; arrow fn
